Move single recipe page endpoint
