+++ tests/test_main/test_route_endpoints.py
@@ -456,80 +456,6 @@ class TestUpdateDiffEndpoint:
 
 
 @pytest.mark.anyio
-class TestGetSingleRecipePage:
-    RECIPE_ID = 123
-    API_URL = f"/api/v0/recipes/{RECIPE_ID}"
-    PAGE_URL = f"/recipes/{RECIPE_ID}"
-
-    @patch("meal_planner.main.internal_client.get")
-    async def test_get_single_recipe_page_api_404(
-        self, mock_api_get, client: AsyncClient
-    ):
-        """Test handling when the API returns 404 for the recipe ID."""
-        mock_api_get.side_effect = httpx.HTTPStatusError(
-            "Not Found",
-            request=httpx.Request("GET", self.API_URL),
-            response=httpx.Response(404),
-        )
-        response = await client.get(self.PAGE_URL)
-        assert response.status_code == 200
-        assert "Recipe Not Found" in response.text
-        mock_api_get.assert_awaited_once_with(self.API_URL)
-
-    @patch("meal_planner.main.internal_client.get")
-    async def test_get_single_recipe_page_api_other_status_error(
-        self, mock_api_get, client: AsyncClient
-    ):
-        """Test handling when the API returns a non-404 status error."""
-        mock_api_get.side_effect = httpx.HTTPStatusError(
-            "Server Error",
-            request=httpx.Request("GET", self.API_URL),
-            response=httpx.Response(500),
-        )
-        response = await client.get(self.PAGE_URL)
-        assert response.status_code == 200
-        assert "Error fetching recipe from API." in response.text
-        assert "Error" in response.text
-        mock_api_get.assert_awaited_once_with(self.API_URL)
-
-    @patch("meal_planner.main.internal_client.get")
-    async def test_get_single_recipe_page_api_generic_error(
-        self, mock_api_get, client: AsyncClient
-    ):
-        """Test handling when the API call raises a generic exception."""
-        mock_api_get.side_effect = Exception("Unexpected API failure")
-        response = await client.get(self.PAGE_URL)
-        assert response.status_code == 200
-        assert "An unexpected error occurred." in response.text
-        mock_api_get.assert_awaited_once_with(self.API_URL)
-
-    RECIPE_ID = 456
-    API_URL = f"/api/v0/recipes/{RECIPE_ID}"
-    PAGE_URL = f"/recipes/{RECIPE_ID}"
-
-    async def test_get_single_recipe_page_success(self, client: AsyncClient):
-        recipe_payload = {
-            "name": "Specific Recipe Page Test",
-            "ingredients": ["Specific Ing 1", "Specific Ing 2"],
-            "instructions": ["Specific Step 1.", "Specific Step 2."],
-        }
-        create_resp = await client.post("/api/v0/recipes", json=recipe_payload)
-        assert create_resp.status_code == 201
-        created_recipe_id = create_resp.json()["id"]
-
-        page_url = f"/recipes/{created_recipe_id}"
-        response = await client.get(page_url)
-        assert response.status_code == 200
-        html_content = response.text
-
-        assert recipe_payload["name"] in html_content
-        assert recipe_payload["ingredients"][0] in html_content
-        assert recipe_payload["ingredients"][1] in html_content
-        assert recipe_payload["instructions"][0] in html_content
-        assert recipe_payload["instructions"][1] in html_content
-
-
-@pytest.mark.anyio
 class TestDeleteRecipeEndpoint:
     DELETE_PATH = "/recipes/delete"
